refactor: extract OpenEditor into internal/editor, out of the TUI package
tui.OpenEditor is exec-based (launches $EDITOR on a temp file) — it has no
Bubble Tea dependency and was only in internal/tui by accident of placement.
Move it to a new lightweight internal/editor package (editor.Open) so callers
that just need an editor don't pull in the heavy TUI package.

Also move the small interactivity helpers it needs — CheckInteractiveAllowed
and ErrInteractiveDisabled — down to internal/utils (next to IsInteractive),
and re-export them from internal/tui so existing tui callers and errors.Is
checks are unchanged (same error var).

Updates all five OpenEditor call sites. create/message and describe now have
no internal/tui import at all (12 -> 10 tui-importing action files);
reorder, submit_metadata, and the split CLI handler keep tui for their
genuine Bubble Tea prompts. Behavior preserved (lint, tui/utils unit tests,
integration 494).
